clean the playlist route
