+++ app/routes/_app.playlists.view.$playlistId/components/playlist-view.tsx
@@ -4,15 +4,12 @@ import { usePlaylistViewData } from "../route";
 
 export function PlaylistView({ isEditing }: { isEditing: boolean }) {
 	const data = usePlaylistViewData();
-	if (data.blocked) {
-		return null;
-	}
 	const { playlistWithGames, userCollection, session } = data;
 	const userCollectionGameIds = userCollection.map((c) => c.gameId);
 
 	return (
 		<LibraryView>
-			{playlistWithGames.games.map((game) => (
+			{playlistWithGames!.games.map((game) => (
 				<div key={game.game.id} className="flex flex-col gap-2">
 					<GameCover coverId={game.game.cover.imageId} gameId={game.gameId} />
 					<PlaylistEntryControls
