+++ app/routes/_app.playlists.view.blocked/route.tsx
@@ -1,0 +1,5 @@
+export default function BlockedPlaylistRoute() {
+  return (
+    <div>This playlist is private</div>
+  )
+}
